[fix] 몬스터 사망 속도 및 보스 애니메이션 일부 변경

+++ Assets/_Scripts/Monster/Core/Base/StateMachineMonster.cs
@@ -70,7 +70,8 @@ public abstract class StateMachineMonster : NormalMonster
         {
             Attack.StopAttack();
             //attacker?.GetComponent<PlayerCharacter>()?.Kill();
-            PlayerCharacter.Instance.Kill();
+            if(PlayerCharacter.Instance != null)
+                PlayerCharacter.Instance.Kill();
             _stateMachine.ChangeState(Game.Monster.StateType.Die);
         }
         else if(!IsSuperArmor)
